requirements.txt
pymongo
PyJWT
flasgger
joblib
marshmallow
wheel
pyopenssl
requests
requests_hawk
Click==7.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
uWSGI==2.0.18

